♻️ rewards-dev | Move command bodies out of Program.cs into Core/Commands.cs
Program.cs is now just arg-parsing + dispatch; the Usage/Fail/ResolveApi/Apply/
Show/Reset implementations live in Core/Commands.cs. Same behaviour — build +
usage/switch/error-path smoke unchanged.
